feat: Move BaseViewTestCase to main app
